requirements.txt
requests~=2.32.4
beautifulsoup4~=4.13.4
python-dotenv~=1.1.1
python-telegram-bot~=22.2
SQLAlchemy~=2.0.38
EbookLib~=0.19
click~=8.2.1
readability-lxml